Re-categorize MessageProtocol packages
