+++ mirai-core/src/commonMain/kotlin/message/protocol/impl/TextProtocol.kt
@@ -15,10 +15,15 @@ import kotlinx.io.core.readBytes
 import kotlinx.io.core.readUInt
 import net.mamoe.mirai.contact.Group
 import net.mamoe.mirai.contact.nameCardOrNick
-import net.mamoe.mirai.internal.message.protocol.*
-import net.mamoe.mirai.internal.message.protocol.MessageEncoderContext.Companion.CONTACT
-import net.mamoe.mirai.internal.message.protocol.MessageEncoderContext.Companion.isForward
-import net.mamoe.mirai.internal.message.protocol.MessageEncoderContext.Companion.originalMessage
+import net.mamoe.mirai.internal.message.protocol.MessageProtocol
+import net.mamoe.mirai.internal.message.protocol.ProcessorCollector
+import net.mamoe.mirai.internal.message.protocol.decode.MessageDecoder
+import net.mamoe.mirai.internal.message.protocol.decode.MessageDecoderContext
+import net.mamoe.mirai.internal.message.protocol.encode.MessageEncoder
+import net.mamoe.mirai.internal.message.protocol.encode.MessageEncoderContext
+import net.mamoe.mirai.internal.message.protocol.encode.MessageEncoderContext.Companion.CONTACT
+import net.mamoe.mirai.internal.message.protocol.encode.MessageEncoderContext.Companion.isForward
+import net.mamoe.mirai.internal.message.protocol.encode.MessageEncoderContext.Companion.originalMessage
 import net.mamoe.mirai.internal.network.protocol.data.proto.ImMsgBody
 import net.mamoe.mirai.message.data.*
 import net.mamoe.mirai.utils.read
